+++ src/main/ts/ephox/alloy/behaviour/toggling/ToggleSchema.ts
@@ -1,4 +1,4 @@
-import { FieldSchema, ValueSchema, FieldProcessorAdt } from '@ephox/boulder';
+import { FieldSchema, ValueSchema } from '@ephox/boulder';
 import { Fun } from '@ephox/katamari';
 
 import * as Fields from '../../data/Fields';
@@ -6,7 +6,7 @@ import * as ToggleModes from './ToggleModes';
 
 export default [
   FieldSchema.defaulted('selected', false),
-  FieldSchema.strict('toggleClass'),
+  FieldSchema.option('toggleClass'),
   FieldSchema.defaulted('toggleOnExecute', true),
 
   FieldSchema.defaultedOf('aria', {